Reorganize files and change Verifier header colour to Amber600
