+++ android/Showcase/src/main/java/com/spruceid/mobilesdkexample/ui/components/HomeHeader.kt
@@ -29,6 +29,7 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.geometry.Size
 import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.graphics.RadialGradientShader
 import androidx.compose.ui.graphics.Shader
 import androidx.compose.ui.graphics.ShaderBrush
 import androidx.compose.ui.graphics.painter.Painter
@@ -41,7 +42,6 @@ import androidx.compose.ui.unit.sp
 import com.spruceid.mobilesdkexample.R
 import com.spruceid.mobilesdkexample.ui.theme.ColorStone950
 import com.spruceid.mobilesdkexample.ui.theme.Switzer
-import com.spruceid.mobilesdkexample.wallet.wallethomeview.createEllipticalGradientShader
 
 data class HeaderButton(
     val icon: Painter,
@@ -163,4 +163,33 @@ fun HomeHeader(
             }
         }
     }
+}
+
+/**
+ * Creates an elliptical radial gradient shader that emanates from the top center
+ * @param radiusYFactor The vertical radius as a factor of height (e.g., 0.95 for 95% of height)
+ */
+fun createEllipticalGradientShader(
+    size: Size,
+    colors: List<Color>,
+    radiusYFactor: Float = 0.95f
+): Shader {
+    val centerX = size.width * 0.5f  // 50% horizontal center
+    val centerY = 0f                 // 0% (top)
+    val radiusX = size.width * 1.0f  // 100% of width
+    val radiusY = size.height * radiusYFactor
+
+    // Create transformation matrix to make ellipse
+    val matrix = android.graphics.Matrix()
+    matrix.preScale(1f, radiusY / radiusX)
+    matrix.preTranslate(0f, -centerY * (radiusX / radiusY - 1f))
+
+    val shader = RadialGradientShader(
+        center = Offset(centerX, centerY * radiusX / radiusY),
+        radius = radiusX,
+        colors = colors,
+        colorStops = listOf(0f, 1f)
+    )
+    shader.setLocalMatrix(matrix)
+    return shader
 }
