+++ android/Showcase/src/main/java/com/spruceid/mobilesdkexample/wallet/WalletHomeView.kt
@@ -1,22 +1,15 @@
 package com.spruceid.mobilesdkexample.wallet
 
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.foundation.verticalScroll
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
@@ -27,13 +20,9 @@ import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.navigation.NavController
 import com.google.accompanist.swiperefresh.SwipeRefresh
 import com.google.accompanist.swiperefresh.rememberSwipeRefreshState
@@ -42,10 +31,11 @@ import com.spruceid.mobilesdkexample.LoadingView
 import com.spruceid.mobilesdkexample.R
 import com.spruceid.mobilesdkexample.db.WalletActivityLogs
 import com.spruceid.mobilesdkexample.navigation.Screen
-import com.spruceid.mobilesdkexample.ui.theme.ColorBase150
-import com.spruceid.mobilesdkexample.ui.theme.ColorStone400
-import com.spruceid.mobilesdkexample.ui.theme.ColorStone950
-import com.spruceid.mobilesdkexample.ui.theme.Inter
+import com.spruceid.mobilesdkexample.ui.components.HeaderButton
+import com.spruceid.mobilesdkexample.ui.components.HomeHeader
+import com.spruceid.mobilesdkexample.ui.theme.ColorAmber600
+import com.spruceid.mobilesdkexample.ui.theme.ColorBase1
+import com.spruceid.mobilesdkexample.ui.theme.ColorBlue600
 import com.spruceid.mobilesdkexample.utils.activityHiltViewModel
 import com.spruceid.mobilesdkexample.utils.credentialDisplaySelector
 import com.spruceid.mobilesdkexample.utils.getCredentialIdTitleAndIssuer
@@ -56,7 +46,6 @@ import com.spruceid.mobilesdkexample.viewmodels.HacApplicationsViewModel
 import com.spruceid.mobilesdkexample.viewmodels.HelpersViewModel
 import com.spruceid.mobilesdkexample.viewmodels.StatusListViewModel
 import com.spruceid.mobilesdkexample.viewmodels.WalletActivityLogsViewModel
-import com.spruceid.mobilesdkexample.wallet.wallethomeview.WalletHomeHeader
 import com.spruceid.mobilesdkexample.walletsettings.generateMockMdl
 import kotlinx.coroutines.launch
 
@@ -83,7 +72,30 @@ fun WalletHomeView(
     }
 }
 
+@Composable
+fun WalletHomeHeader(
+    navController: NavController?
+) {
+    val gradientColors = listOf(ColorBlue600, ColorBase1)
+    val buttons = listOf(
+        HeaderButton(
+            icon = painterResource(id = R.drawable.qrcode_scanner),
+            contentDescription = stringResource(id = R.string.qrcode_scanner),
+            onClick = { navController?.navigate(Screen.ScanQRScreen.route) }
+        ),
+        HeaderButton(
+            icon = painterResource(id = R.drawable.user),
+            contentDescription = stringResource(id = R.string.user),
+            onClick = { navController?.navigate(Screen.WalletSettingsHomeScreen.route) }
+        )
+    )
 
+    HomeHeader(
+        title = "Wallet",
+        gradientColors = gradientColors,
+        buttons = buttons
+    )
+}
 
 @Composable
 fun WalletHomeBody(
